feat: fast-fiedler pipeline, unified labeling, target-layer sync
Performance — spectral bottleneck eliminated:
- Add --fast-fiedler flag to `audit` subcommand. Single forward pass
  with output_attentions=True; slices attention to tool-call token span
  [t_func:t_end] or last-50-token fallback; runs GPU Lanczos
  (DirectedTopologist, k=20) on the S=50 subgraph instead of full N=N
  dense eigh. Benchmarked: 177x speedup on 468-token ToolBench records
  (19 ms vs 3359 ms per-sample spectral step). Worst-case 12-minute
  spikes now complete in ~8 s end-to-end.
- Import DirectedTopologist and calculate_spectral_velocity from
  spectral_trust 0.2.1; propagate max_velocity into layer_diagnostics.

Research integrity — labeling unified:
- Rewrite spectral_guardrails/probes/labeling.py assign_label() with
  the full JSON-aware schema bridge (previously only in scratch/eval_baselines.py).
  Handles OpenAI-style {name, parameters}, BFCL GT-style {fn_name: {arg: [val]}},
  and Gorilla-style schemas. Adds numeric tolerance (1e-6), string normalization,
  parallel call matching, and ToolBench prose fallback. Training extraction
  and production audit now produce identical labels on the same data.

Dimension correctness — probe vector sync:
- Align target_layers in handle_extract with handle_audit:
  [4, 8, 16, 20, 24, 28, 31] (7 layers, 86016-dim) corrected to
  [4, 8, 11, 15, 20, 24, 28, 31] (8 layers, 98304-dim). Probe trained
  on extraction data is now dimension-compatible with audit features.

Dead code removal:
- Delete ~145 lines of orphaned module-level code after
  `if __name__ == "__main__": main()` that referenced undefined names
  (checkpoint_path, args, config) causing NameError on every CLI run.

Additional:
- Add Qwen3.5-2B to overnight run schedule (run_overnight.py).
- BnB 4-bit compute dtype float16 (was bfloat16) for stable quantized
  inference on hardware without native BF16 support (models.py).
- Extended probe feature extraction and data normalization utilities.

+++ spectral_guardrails/utils/models.py
@@ -47,7 +47,7 @@ def load_model_and_tokenizer(model_key_or_path: str,
     if load_in_4bit:
         quantization_config = BitsAndBytesConfig(
             load_in_4bit=True,
-            bnb_4bit_compute_dtype=torch.bfloat16,
+            bnb_4bit_compute_dtype=torch.float16,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type="nf4"
         )
